refactor(): move discussionSearchChannels helpers
